+++ internal/app/balance_service_app/handlers/account_handler/request_models/transfer_message.go
@@ -1,8 +1,8 @@
 package request_models
 
 type TransferMessage struct {
-	SrcUserID int64   `json:"src_user_id"`
-	DstUserID int64   `json:"dst_user_id"`
+	SrcUserID int     `json:"src_user_id"`
+	DstUserID int     `json:"dst_user_id"`
 	Sum       float64 `json:"sum"`
 	Comment   string  `json:"comment"`
 }
