+++ internal/app/balance_service_app/handlers/service_handler/request_models/buy_service_message.go
@@ -1,9 +1,9 @@
 package request_models
 
 type BuyServiceMessage struct {
-	UserID    int64   `json:"user_id"`
-	OrderID   int64   `json:"order_id"`
-	ServiceID int64   `json:"service_id"`
+	UserID    int     `json:"user_id"`
+	OrderID   int     `json:"order_id"`
+	ServiceID int     `json:"service_id"`
 	Sum       float64 `json:"sum"`
 	Comment   string  `json:"comment"`
 }
